pyproject.toml
[project]
name = "drise-dmfpp-ddeletion"
version = "0.1.0"
description = ""
authors = [
    {name = "Linda Riesen",email = "rieselin@students.zhaw.ch"}
]
readme = "README.md"
requires-python = "3.10"
dependencies = [
    "accelerate (==1.10.1)",
    "aiohappyeyeballs (==2.6.1)",
    "aiohttp (==3.13.0)",
    "aiosignal (==1.4.0)",
    "annotated-types (==0.7.0)",
    "anyio (==4.11.0)",
    "asttokens (==3.0.0)",
    "async-timeout (==5.0.1)",
    "attrs (==25.4.0)",
    "bitsandbytes (==0.48.1)",
    "certifi (==2025.10.5)",
    "charset-normalizer (==3.4.3)",
    "colorama (==0.4.6)",
    "comm (==0.2.3)",
    "contourpy (==1.3.2)",
    "cut-cross-entropy (==25.1.1)",
    "cycler (==0.12.1)",
    "datasets (==4.2.0)",
    "debugpy (==1.8.17)",
    "decorator (==5.2.1)",
    "diffusers (==0.35.2)",
    "dill (==0.4.0)",
    "docstring-parser (==0.17.0)",
    "exceptiongroup (==1.3.0)",
    "executing (==2.2.1)",
    "filelock (==3.20.0)",
    "fonttools (==4.60.1)",
    "frozenlist (==1.8.0)",
    "fsspec (==2025.9.0)",
    "h11 (==0.16.0)",
    "hf-transfer (==0.1.9)",
    "httpcore (==1.0.9)",
    "httpx (==0.28.1)",
    "huggingface-hub (==0.35.3)",
    "idna (==3.10)",
    "imageio (==2.37.0)",
    "importlib-metadata (==8.7.0)",
    "ipykernel (==6.30.1)",
    "ipython (==8.37.0)",
    "jedi (==0.19.2)",
    "jinja2 (==3.1.6)",
    "joblib (==1.5.2)",
    "jsonschema (==4.25.1)",
    "jsonschema-specifications (==2025.9.1)",
    "jupyter-client (==8.6.3)",
    "jupyter-core (==5.8.1)",
    "kiwisolver (==1.4.9)",
    "lazy-loader (==0.4)",
    "lime (==0.2.0.1)",
    "markdown-it-py (==4.0.0)",
    "markupsafe (==3.0.3)",
    "matplotlib (==3.10.7)",
    "matplotlib-inline (==0.1.7)",
    "mdurl (==0.1.2)",
    "mistral-common (==1.8.5)",
    "mpmath (==1.3.0)",
    "msgspec (==0.19.0)",
    "multidict (==6.7.0)",
    "multiprocess (==0.70.16)",
    "nest-asyncio (==1.6.0)",
    "networkx (==3.4.2)",
    "numpy (==1.26.4)",
    "opencv-python (==4.11.0.86)",
    "packaging (==25.0)",
    "pandas (==2.3.3)",
    "parso (==0.8.5)",
    "peft (==0.17.1)",
    "pillow (==11.3.0)",
    "platformdirs (==4.5.0)",
    "polars (==1.34.0)",
    "polars-runtime-32 (==1.34.0)",
    "prompt-toolkit (==3.0.52)",
    "propcache (==0.4.1)",
    "protobuf (==6.32.1)",
    "psutil (==7.1.0)",
    "pure-eval (==0.2.3)",
    "pyarrow (==21.0.0)",
    "pycountry (==24.6.1)",
    "pydantic (==2.12.2)",
    "pydantic-extra-types (==2.10.6)",
    "pydantic-core (==2.41.4)",
    "pygments (==2.19.2)",
    "pyparsing (==3.2.5)",
    "python-dateutil (==2.9.0.post0)",
    "pytz (==2025.2)",
    "pywin32 (==311)",
    "pyyaml (==6.0.3)",
    "pyzmq (==27.1.0)",
    "referencing (==0.37.0)",
    "regex (==2025.9.18)",
    "requests (==2.32.5)",
    "rich (==14.2.0)",
    "rpds-py (==0.27.1)",
    "safetensors (==0.6.2)",
    "scikit-image (==0.25.2)",
    "scikit-learn (==1.7.2)",
    "scipy (==1.15.3)",
    "sentencepiece (==0.2.1)",
    "shapely (==2.1.2)",
    "shtab (==1.7.2)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "stack-data (==0.6.3)",
    "sympy (==1.14.0)",
    "threadpoolctl (==3.6.0)",
    "tifffile (==2025.5.10)",
    "tiktoken (==0.12.0)",
    "tokenizers (==0.22.1)",
    "torch==2.8.0+cu126",
    "torchaudio==2.8.0+cu126",
    "torchvision==0.23.0+cu126",
    "torchao==0.14.0",
    "tornado (==6.5.2)",
    "tqdm (==4.67.1)",
    "traitlets (==5.14.3)",
    "transformers (==4.56.1)",
    "triton-windows (==3.4.0.post20)",
    "trl (==0.22.2)",
    "typeguard (==4.4.4)",
    "typing-inspection (==0.4.2)",
    "typing-extensions (==4.15.0)",
    "tyro (==0.9.35)",
    "tzdata (==2025.2)",
    "ultralytics (==8.3.209)",
    "ultralytics-thop (==2.0.17)",
    "unsloth (==2025.10.3)",
    "unsloth-zoo (==2025.10.3)",
    "urllib3 (==2.5.0)",
    "wcwidth (==0.2.14)",
    "xformers (==0.0.32.post2)",
    "xxhash (==3.6.0)",
    "yarl (==1.22.0)",
    "zipp (==3.23.0)",
    "fiftyone (>=1.9.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu126"
priority = "supplemental"


[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu126"
priority = "supplemental"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.poetry]
package-mode = false
